Update terraform deps to 0.11.2
Why it was needed:
- To be able to run acceptance test for import we needed to have a newer version of terraform (0.10 was introducing a bug on import test helper)
- That's never a bad idea to build provider with newer terraform

Having terraform dependencies in vendor make things painful when we want to update it, maybe ignoring `github.com/hashicorp/terraform` in vendor could be a good idea, we could set a version when building it. That's actually what we did on https://github.com/orange-cloudfoundry/terraform-provider-cloudfoundry .
What do you thing about this @mevansam ?

This version add a check on field name and this check was complaining about fields `syslogDrainURL` and `routeServiceURL` in resource user provided service, example of message:
```
Step 0 error: config is invalid: provider.cf: Internal validation of the provider failed! This is always a bug
with the provider itself, and not a user issue. Please report
this bug:

1 error(s) occurred:

* resource cf_user_provided_service: syslogDrainURL: Field name may only contain lowercase alphanumeric characters & underscores.
```

So I've marked those 2 fields as deprecated and introduced `syslog_drain_url` and `route_service_url` and modified the doc and tests consequently. This will have no impact on previous deployment.