stop downloading the app asynchronously
The problem is that the schema.ResourceData is not thread safe, even
for concurrent reads, due to internal memoization during reads.  The
result was that app creation would sometimes fail mysteriously with an
"unexpected EOF" error due to the provider process crashing due to
either concurrent map writes or read/write.