Rework cf_app local download to use ioutil.Temp & clean up temp files #57
@mevansam could you do code review please ?

I've removed inside `repo_manager.GetGitRepository` the ability to get back an already git cloned repository as temp directory can't be took again.
I think it's not that much useful as we run terraform on a pipeline so git cloned is anyway always performed. Also, issue #58 address this problem to not download for nothing.

About the cleaning, the way i fix it is definetely not the best way to do so but it seems that I will need to refactor `repo_manager`. Do you share this feeling @mevansam ?