improve handling of cf_app binary updates
Cloud Foundry marks an application's package_state as pending
whenever a a new binary is uploaded.  This leads to calls to the
restage endpoint actually throwing an error instead of prompting
immediate restaging.  The CF CLI actually forces an immediate restage
by restarting the application.  Thus this is what we will now do.