Revert "improve app import for routes"
In practice, this additional "helper" update to the state proved detrimental.
I believe the normal case when importing an application would be that
its route(s) would be imported as well.  In this case, depending on how
the route is mapped to the application, either the route will be unmapped by
the cf_route resource and then remapped by the cf_app resource, or
no actual update will happen.  Either way, this change interfered with
that process.

This reverts commit 6e4355aa0a0c75fa48d84e7ee23fa5a54c0c23eb.