refactor and improve how restaging/restarting is handled
In order to increase the readability and improve the stability of the code
the structure of resourceAppUpdate has been revisited in its entirety.
The decisions around when to restart/restage an application should
be far more clear now as well as being more robust for any possible edge
cases.

For more detail see: https://github.com/mevansam/terraform-provider-cf/pull/79