Implement SSO passcode-based authentication flow
Fixes #226

This makes the provider compatible with Cloud Foundry distributions which
use single sign on (SSO) instead of username/password to authenticate
users.

This adds a new environment variable `CF_SSO_PASSCODE` which takes the
same value as `cf login --sso-passcode <passcode>`. This also implements
persisting and refreshing the access token to disk to avoid the need for
repeated login attempts, which are incompatible with the SSO flow.