Fallback check in case OrgManager has fewer permissions
It looks like newer version of the CF API limit
capabilities of OrgManagers. The call to get an 
overview of a user’s orgs could fail. In this case we 
can fall back to a (slower) check by looping through
all current users of the Org. This allows the operation
To proceed. This is crucial of the provider is 
operated by OrgManager accounts as it would otherwise
not allow them to create and populate spaces.